feat: Implement debug commands for vector search testing
- Added `debugCommands.ts` to register and handle vector search test commands.
- Integrated user input for search queries and displayed results in the debug console.
- Implemented error handling and user notifications for service initialization issues.

feat: Create PolarisService for documentation processing

- Introduced `PolarisService.ts` to manage documentation processing, symbol parsing, and AI documentation generation.
- Added performance monitoring and error handling throughout the processing workflow.
- Implemented embedding reconciliation for symbols with docstrings.

feat: Develop VscodeService for user notifications

- Created `VscodeService.ts` to abstract VSCode notification APIs for improved testability.
- Implemented methods for showing information, error, and warning messages to users.

refactor: Centralize service initialization logic

- Created `index.ts` to manage the initialization of embedding and vector storage services.
- Added comprehensive error handling and logging for service initialization metrics.
- Implemented graceful degradation for the extension in case of service failures.
