--- tsconfig.test.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "target": "ES2022",
-    "outDir": "out/test",
-    "lib": [
-      "ES2022"
-    ],
-    "sourceMap": true,
-    "rootDir": "src",
-    "strict": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "skipLibCheck": true
-  },
-  "include": [
-    "src/test/**/*.ts"
-  ],
-  "exclude": [
-    "vite.config.ts",
-    "vite.config.js",
-    "*.config.ts",
-    "*.config.js"
-  ]
-}
